+++ notifiers/slack_test.go
@@ -1,0 +1,72 @@
+// Statping
+// Copyright (C) 2018.  Hunter Long and the project contributors
+// Written by Hunter Long <info@socialeck.com> and the project contributors
+//
+// https://github.com/hunterlong/statping
+//
+// The licenses for most software and other practical works are designed
+// to take away your freedom to share and change the works.  By contrast,
+// the GNU General Public License is intended to guarantee your freedom to
+// share and change all versions of a program--to make sure it remains free
+// software for all its users.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+package notifiers
+
+import (
+	"github.com/statping/statping/database"
+	"github.com/statping/statping/types/notifications"
+	"github.com/statping/statping/types/null"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+	"os"
+	"testing"
+	"time"
+)
+
+var (
+	SLACK_URL string
+)
+
+func TestSlackNotifier(t *testing.T) {
+	db, err := database.OpenTester()
+	require.Nil(t, err)
+	db.AutoMigrate(&notifications.Notification{})
+	notifications.SetDB(db)
+
+	SLACK_URL = os.Getenv("SLACK_URL")
+	slacker.Host = SLACK_URL
+	slacker.Enabled = null.NewNullBool(true)
+
+	if SLACK_URL == "" {
+		t.Log("slack notifier testing skipped, missing SLACK_URL environment variable")
+		t.SkipNow()
+	}
+
+	t.Run("Load slack", func(t *testing.T) {
+		slacker.Host = SLACK_URL
+		slacker.Delay = time.Duration(100 * time.Millisecond)
+		slacker.Limits = 3
+		Add(slacker)
+		assert.Equal(t, "Hunter Long", slacker.Author)
+		assert.Equal(t, SLACK_URL, slacker.Host)
+	})
+
+	t.Run("slack Within Limits", func(t *testing.T) {
+		ok := slacker.CanSend()
+		assert.True(t, ok)
+	})
+
+	t.Run("slack OnFailure", func(t *testing.T) {
+		err := slacker.OnFailure(exampleService, exampleFailure)
+		assert.Nil(t, err)
+	})
+
+	t.Run("slack OnSuccess", func(t *testing.T) {
+		err := slacker.OnSuccess(exampleService)
+		assert.Nil(t, err)
+	})
+
+}