+++ types/notifications/methods.go
@@ -64,7 +64,7 @@ func (n *Notification) CanSend() bool {
 	//fmt.Println("Last sent before now: ", n.lastSent.Add(60*time.Second).Before(utils.Now()))
 
 	// the last sent notification was past 1 minute (limit per minute)
-	if n.lastSent.Add(60 * time.Second).Before(utils.Now()) {
+	if n.lastSent.Add(60 * time.Minute).Before(utils.Now()) {
 		if n.lastSentCount != 0 {
 			n.lastSentCount--
 		}