+++ types/services/notifier.go
@@ -13,6 +13,16 @@ func AllNotifiers() []ServiceNotifier {
 	return allNotifiers
 }
 
+func FindNotifier(method string) *notifications.Notification {
+	for _, n := range allNotifiers {
+		notif := n.Select()
+		if notif.Method == method {
+			return notif
+		}
+	}
+	return nil
+}
+
 type ServiceNotifier interface {
 	OnSuccess(*Service) error                    // OnSuccess is triggered when a service is successful
 	OnFailure(*Service, *failures.Failure) error // OnFailure is triggered when a service is failing